video: rockchip: mpp：fix reset_group protected range
In some platform, there are some devices share PD.
If one device is doing pmu idle request to cru reset and
at this time cpu want to read/write reg for the device will crash.

So use the reset_group to prevent the case.

There is a issue that reset_group protected range not enough,
so fix it.

Signed-off-by: Yandong Lin <yandong.lin@rock-chips.com>
Change-Id: Ide32966f7cff842e4611213d26901617fd57bc14